requirements.txt
# Flask & Gunicorn
Flask==1.1.2
gunicorn==20.0.4
Flask-Cors==3.0.9

# Google API
google-api-python-client==1.12.3
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
oauth2client==4.1.3